--- app/src/main/java/com/example/cooktifyapp/view/navigation/inputImages/InputFragment.kt
@@ -1,81 +1,0 @@
-
-import android.Manifest
-import android.content.pm.PackageManager
-import android.net.Uri
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.activity.result.PickVisualMediaRequest
-import androidx.activity.result.contract.ActivityResultContracts
-import androidx.core.content.ContextCompat
-import androidx.fragment.app.Fragment
-import com.example.cooktifyapp.databinding.FragmentInputBinding
-
-
-class InputFragment : Fragment() {
-
-    private lateinit var binding: FragmentInputBinding
-    private var currentImageUri: Uri? = null
-
-    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        binding = FragmentInputBinding.inflate(inflater,container,false)
-        val root = binding.root
-        return root
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        binding.ivgaleri.setOnClickListener {
-            startGalery()
-        }
-    }
-    companion object {
-        private const val REQUIRED_PERMISSION = Manifest.permission.CAMERA
-    }
-    private fun allPermissionGranted() =
-        ContextCompat.checkSelfPermission(
-            requireContext(),
-            REQUIRED_PERMISSION
-        ) == PackageManager.PERMISSION_GRANTED
-
-    private val requestPermissionLauncher =
-        registerForActivityResult(
-            ActivityResultContracts.RequestPermission()
-        ) { isGranted: Boolean ->
-            if (isGranted) {
-                startGalery()
-            } else {
-
-            }
-        }
-
-    private fun startGalery(){
-        if (allPermissionGranted()) {
-            launcherGallery.launch(PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly))
-        } else {
-            requestPermissionLauncher.launch(REQUIRED_PERMISSION)
-        }
-    }
-    private val launcherGallery = registerForActivityResult(
-        ActivityResultContracts.PickVisualMedia()
-    ) { uri: Uri? ->
-        if (uri != null) {
-            currentImageUri = uri
-            showImage()
-        } else {
-
-        }
-    }
-    private fun showImage(){
-        currentImageUri?.let {uri ->
-            binding.itemsImg.setImageURI(uri)
-        }
-
-    }
-
-
-
-
-}
